+++ benchmarks/perf-test/src/main.rs
@@ -28,7 +28,7 @@ enum MeasureMode {
 #[derive(Debug, Parser)]
 #[clap(name = "cervo perf-tests")]
 struct RustyPerf {
-    #[clap(subcommand)] // Note that we mark a field as a subcommand
+    #[clap(subcommand)]
     mode: MeasureMode,
 }
 
@@ -43,7 +43,7 @@ struct RustyPerf {
 //         let start = Instant::now();
 
 //         for _ in 0..10 {
-//             instance.infer(observations.clone()).unwrap();
+//             instance.infer_batch(observations.clone()).unwrap();
 //         }
 //         let elapsed = start.elapsed();